feat: componentizando a aplicação
