+++ ignite-reactjs-2023/03-ignite-timer/src/pages/Home/index.tsx
@@ -12,6 +12,8 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import * as zod from "zod";
 import { useEffect, useState } from "react";
 import { differenceInSeconds } from "date-fns";
+import { NewCycleForm } from "./components/NewCycleForm";
+import { Countdown } from "./components/Countdown";
 
 const newCycleFormValidationSchema = zod.object({
     task: zod.string().min(1, "Informe a tarefa"),
@@ -136,46 +138,9 @@ export function Home(){
     return (
         <HomeContainer>
             <form onSubmit={handleSubmit(handleCreateNewCycle)} action="">
-                <FormContainer>
-                    <label htmlFor="">Vou trabalhar em</label>
-                    <TaskInput
-                        id="task"
-                        list="task-suggestions"
-                        placeholder="Dê um nome para o seu projeto"
-                        disabled = {!!activeCycle}
-                        {...register("task")}
-                    />
+                <NewCycleForm />
+                <Countdown />
 
-                    <datalist id="task-suggestions">
-                        <option value="Projeto 1" />
-                        <option value="Projeto 2" />
-                        <option value="Projeto 3" />
-                        <option value="Maçã" />
-                    </datalist>
-
-                    <label htmlFor="">durante</label>
-                    <MinutesAmountInput
-                        type="number"
-                        id="minutesAmount"
-                        placeholder="00"
-                        step={5}
-                        min={1}
-                        max={60}
-                        disabled = {!!activeCycle}
-                        {...register("minutesAmount", { valueAsNumber: true })}
-                    />
-
-                    <span>minutos.</span>
-                </FormContainer>
-
-                <CountdownContainer>
-                    <span>{minutes[0]}</span>
-                    <span>{minutes[1]}</span>
-                    <Separator>:</Separator>
-                    <span>{seconds[0]}</span>
-                    <span>{seconds[1]}</span>
-                </CountdownContainer>
-                
                 { activeCycle ?  (
                     <StopCountdownButton onClick={handleInterruptCycle} type="submit">
                         <HandPalm size={24} />
